requirements.txt
flask
flask-cors
python-dotenv
pymongo
pandas
streamlit
plotly
pycoingecko